refactor: Move node-listing route to 'vp' app
Prevents another special case in the middleware, but also makes sense regarding separation. Installed nodes belong to the 'vp' app overall, not a specific workflow.
