+++ vp/vp/views.py
@@ -1,6 +1,8 @@
 from django.http import JsonResponse
 from rest_framework.decorators import api_view
 from drf_yasg.utils import swagger_auto_schema
+from pyworkflow import Node
+
 
 @swagger_auto_schema(method='get', responses={200:'JSON response with data'})
 @api_view(['GET'])
@@ -18,4 +20,42 @@ def info(request):
         "version": "negative something",
         "about": "super-duper workflows!"
     }
+    return JsonResponse(data)
+
+@swagger_auto_schema(method='get',
+                     operation_summary='Retrieve a list of installed Nodes',
+                     operation_description='Retrieves a list of installed Nodes, in JSON.',
+                     responses={
+                         200: 'List of installed Nodes, in JSON',
+                     })
+@api_view(['GET'])
+def retrieve_nodes_for_user(request):
+    """Assembles list of Nodes accessible to workflows.
+
+    Retrieve a list of classes from the Node module in `pyworkflow`.
+    List is split into 'types' (e.g., 'IO' and 'Manipulation') and
+    'keys', or individual command Nodes (e.g., 'ReadCsv', 'Pivot').
+    """
+    data = dict()
+
+    # Iterate through node 'types'
+    for parent in Node.__subclasses__():
+        data[parent.__name__] = list()
+
+        # Iterate through node 'keys'
+        for child in parent.__subclasses__():
+            # TODO: check attribute-scope is handled correctly
+            child_node = {
+                'name': child.name,
+                'node_key': child.__name__,
+                'node_type': parent.__name__,
+                'num_in': child.num_in,
+                'num_out': child.num_out,
+                'color': child.color or parent.color,
+                'doc': child.__doc__,
+                'options': {**parent.DEFAULT_OPTIONS, **child.DEFAULT_OPTIONS},
+            }
+
+            data[parent.__name__].append(child_node)
+
     return JsonResponse(data)
